requirements.txt
awscli==1.16.258
botocore==1.12.248
certifi==2019.6.16
chardet==3.0.4
idna==2.9
jmespath==0.9.4
Naked==0.1.31
pyasn1==0.4.7
pycryptodome==3.9.7
PyYAML==5.3.1
requests==2.23.0
rsa==3.4.2
s3transfer==0.2.1
shellescape==3.8.1
urllib3==1.25.9
